:twisted_rightwards_arrows: :: [#781] - 도메인 연결시 HTTP 소켓 연결이 끊기는 현상

+++ src/main/kotlin/com/dcd/server/core/domain/domain/usecase/DisconnectDomainUseCase.kt
@@ -3,8 +3,8 @@ package com.dcd.server.core.domain.domain.usecase
 import com.dcd.server.core.common.annotation.UseCase
 import com.dcd.server.core.common.data.WorkspaceInfo
 import com.dcd.server.core.domain.domain.exception.DomainNotFoundException
-import com.dcd.server.core.domain.domain.service.RebootNginxService
-import com.dcd.server.core.domain.domain.service.RemoveHttpConfigService
+import com.dcd.server.core.domain.domain.exception.DomainNotConnectedException
+import com.dcd.server.core.domain.domain.service.*
 import com.dcd.server.core.domain.domain.spi.CommandDomainPort
 import com.dcd.server.core.domain.domain.spi.QueryDomainPort
 
@@ -13,7 +13,8 @@ class DisconnectDomainUseCase(
     private val queryDomainPort: QueryDomainPort,
     private val commandDomainPort: CommandDomainPort,
     private val removeHttpConfigService: RemoveHttpConfigService,
-    private val rebootNginxService: RebootNginxService,
+    private val applyHttpConfigService: ApplyHttpConfigService,
+    private val generateHttpConfigService: GenerateHttpConfigService,
     private val workspaceInfo: WorkspaceInfo
 ) {
     fun execute(domainId: String) {
@@ -23,10 +24,20 @@ class DisconnectDomainUseCase(
         if (workspaceInfo.workspace != domain.workspace)
             throw DomainNotFoundException()
 
+        if (domain.application == null)
+            throw DomainNotConnectedException()
+
         val updatedDomain = domain.copy(application = null)
         commandDomainPort.save(updatedDomain)
 
         removeHttpConfigService.removeHttpConfig(domain)
-        rebootNginxService.rebootNginx()
+        try {
+            applyHttpConfigService.applyHttpConfig()
+        } catch (e: Exception) {
+            // HTTP config 적용 실패 시 도메인 연결 해제후 설정 롤백
+            generateHttpConfigService.generateWebServerConfig(domain.application, domain)
+            commandDomainPort.save(domain)
+            throw e
+        }
     }
 }
